test: move helpers into test/helpers/
