+++ test/helpers/assertStubbedCalls.js
