+++ test/helpers/exitWithCode.js
