+++ test/helpers/fakeChild.js
