+++ test/helpers/initFixture.js
@@ -4,8 +4,13 @@ import path from "path";
 import fse from "fs-extra";
 
 const tmpDir = path.resolve(__dirname, "../tmp");
+const originalCwd = process.cwd();
 
 const createdDirectories = [];
+
+afterEach(() => {
+  process.chdir(originalCwd);
+});
 
 afterAll(() => {
   createdDirectories.map((dir) => rimraf.sync(dir));
@@ -13,17 +18,16 @@ afterAll(() => {
 
 let uniqueId = 0;
 
-export default function initExternalFixture(fixturePath, callback) {
-  const fixtureDir = path.resolve(__dirname, "./fixtures/" + fixturePath);
-  const testDir = path.resolve(tmpDir, "test-external-" + Date.now() + "-" + (uniqueId++));
+export default function initFixture(fixturePath, callback) {
+  const fixtureDir = path.resolve(__dirname, "../fixtures/" + fixturePath);
+  const testDir = path.resolve(tmpDir, "test-" + Date.now() + "-" + (uniqueId++));
 
   createdDirectories.push(testDir);
 
   fse.copy(fixtureDir, testDir, (err) => {
     if (err) return callback(err);
-    child.execSync("git init . && git add -A && git commit -m \"Init external commit\"", {
-      cwd: testDir
-    });
+    process.chdir(testDir);
+    child.execSync("git init . && git add -A && git commit -m \"Init commit\"");
     callback();
   });
