+++ test/helpers/stub.js
